+++ react/src/modules/todo/actions.ts
@@ -1,0 +1,10 @@
+import { createAction } from 'typesafe-actions';
+
+export const ADD_TODO = 'todos/ADD_TODO';
+export const TOGGLE_TODO = 'todos/TOGGLE_TODO';
+export const REMOVE_TODO = 'todos/REMOVE_TODO';
+
+export const addTodo = createAction(ADD_TODO)<string>();
+export const toggleTodo = createAction(TOGGLE_TODO)<number>();
+export const removeTodo = createAction(REMOVE_TODO)<number>();
+
