+++ react/src/modules/todo/reducer.ts
@@ -1,27 +1,6 @@
-import {
-  createAction,
-  ActionType,
-  createReducer
-} from 'typesafe-actions';
-
-const ADD_TODO = 'todos/ADD_TODO' as const;
-const TOGGLE_TODO = 'todos/TOGGLE_TODO' as const;
-const REMOVE_TODO = 'todos/REMOVE_TODO' as const;
-
-export const addTodo = createAction(ADD_TODO)<string>();
-export const toggleTodo = createAction(TOGGLE_TODO)<number>();
-export const removeTodo = createAction(REMOVE_TODO)<number>();
-
-const actions = {addTodo, toggleTodo, removeTodo};
-type TodosAction = ActionType<typeof actions>;
-
-export interface Todo {
-  id: number;
-  text: string;
-  done: boolean;
-}
-
-type TodosState = Todo[];
+import {createReducer} from 'typesafe-actions';
+import { ADD_TODO, REMOVE_TODO, TOGGLE_TODO } from './actions';
+import {TodosState, TodosAction} from './types';
 
 const initialState: TodosState = [
   { id: 1, text: '타입스크립트 배우기', done: true },
