+++ react/src/modules/todo/types.ts
@@ -1,0 +1,12 @@
+import { ActionType } from 'typesafe-actions';
+import * as actions from './actions';
+
+export type TodosAction = ActionType<typeof actions>;
+
+export interface Todo {
+  id: number;
+  text: string;
+  done: boolean;
+}
+
+export type TodosState = Todo[];
